Output numbers in a culturally-independent format (#3)
By default, calling ToString() on an integer or floating-point type
produces a result dependent on the culture settings of the operating
system, which influences properties like the decimal symbol (point or
comma). Since we are outputting source code, force the usage of the
invariant culture which is independent of operating system settings and
matches the expectations of IBIze.